Mock useAuth0 hook with an empty context
Auth0 React wrapper (react-auth0-spa.js) is initializing the Auth0Context without a value
because it will be assigned on rendering.

Turns out that tests does not behave like the browser regarding the React Context and Hooks,
so in that case it returns nothing when called from a component rendered by the testsuite.

In order to avoid a null pointer exceptionso in those situations,

   "TypeError: Cannot read property 'isAuthenticated' of undefined"

it was needed to mock the content returned by useAuth0 in order to get some value from there