Fix grouping of Pull Request Size review chart
This Chart shows PRs in two groups: waiting for review or not.
Grouping criteria should be having review_happened or approve_happened or changes_request_happened or merge_happened or being closed.
But API since the tooltip also needs to show the time waiting for being reviewed, and the events
above do not expose it, we can only differentiate between being review-complete, or not.

Anyhow, having review_requested is not a valid grouping key because:
- a PR can be approved (thus, having review) without having review_requested,
- and a PR can be still waiting for review even having review_requested.

Signed-off-by: David Pordomingo <David.Pordomingo.F@gmail.com>